Issue #2824610 by quietone, Jo Fitzgerald, ohthehugemanatee, heddn, catch: Rename DedupeBase/DedupeEntity process plugins to MakeUnique and add documentation
(cherry picked from commit b2071808eb469137e3b6a82cc7359c84e56693b3)
