+++ core/modules/block/migration_templates/d6_block.yml
@@ -13,7 +13,8 @@ process:
     source: status
   id:
     # We need something unique, so aggregator, aggregator_1 etc will do.
-    plugin: dedupe_entity
+    plugin: make_unique_entity_field
+
     entity_type: block
     field: id
     postfix: _
