+++ core/modules/contact/migration_templates/contact_category.yml
@@ -11,7 +11,7 @@ process:
       plugin: machine_name
       source: category
     -
-      plugin: dedupe_entity
+      plugin: make_unique_entity_field
       entity_type: contact_form
       field: id
       length: 32
