+++ core/modules/migrate/src/Plugin/migrate/process/DedupeBase.php
@@ -2,11 +2,8 @@
 
 namespace Drupal\migrate\Plugin\migrate\process;
 
-use Drupal\migrate\ProcessPluginBase;
-use Drupal\migrate\MigrateExecutableInterface;
-use Drupal\migrate\Row;
-use Drupal\migrate\MigrateException;
-use Drupal\Component\Utility\Unicode;
+@trigger_error('The ' . __NAMESPACE__ . ' \DedupeEntityBase is deprecated in
+Drupal 8.4.x and will be removed before Drupal 9.0.0. Instead, use ' . __NAMESPACE__ . ' \MakeUniqueEntityFieldBase', E_USER_DEPRECATED);
 
 /**
  * This abstract base contains the dedupe logic.
@@ -17,41 +14,9 @@ use Drupal\Component\Utility\Unicode;
  * and incremented until a unique value is created.
  *
  * @link https://www.drupal.org/node/2345929 Online handbook documentation for dedupebase process plugin @endlink
+ *
+ * @deprecated in Drupal 8.4.x and will be removed in Drupal 9.0.x. Use
+ *   \Drupal\migrate\Plugin\migrate\process\MakeUniqueBase instead.
  */
-abstract class DedupeBase extends ProcessPluginBase {
-
-  /**
-   * {@inheritdoc}
-   */
-  public function transform($value, MigrateExecutableInterface $migrate_executable, Row $row, $destination_property) {
-    $i = 1;
-    $postfix = isset($this->configuration['postfix']) ? $this->configuration['postfix'] : '';
-    $start = isset($this->configuration['start']) ? $this->configuration['start'] : 0;
-    if (!is_int($start)) {
-      throw new MigrateException('The start position configuration key should be an integer. Omit this key to capture from the beginning of the string.');
-    }
-    $length = isset($this->configuration['length']) ? $this->configuration['length'] : NULL;
-    if (!is_null($length) && !is_int($length)) {
-      throw new MigrateException('The character length configuration key should be an integer. Omit this key to capture the entire string.');
-    }
-    // Use optional start or length to return a portion of deduplicated value.
-    $value = Unicode::substr($value, $start, $length);
-    $new_value = $value;
-    while ($this->exists($new_value)) {
-      $new_value = $value . $postfix . $i++;
-    }
-    return $new_value;
-  }
-
-  /**
-   * This is a query checking the existence of some value.
-   *
-   * @param mixed $value
-   *   The value to check.
-   *
-   * @return bool
-   *   TRUE if the value exists.
-   */
-  abstract protected function exists($value);
-
+abstract class DedupeBase extends MakeUniqueBase {
 }
